# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API spec for DCFB API.",
    "version" : "0.0.1",
    "title" : "DCFB API"
  },
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
    "/items" : {
      "get" : {
        "tags" : [ "Items" ],
        "summary" : "List items",
        "description" : "List items",
        "operationId" : "listItems",
        "parameters" : [ {
          "name" : "categoryIds",
          "in" : "query",
          "description" : "Filter by category ids (comma delimitered)",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Search by free-text query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "An array of items",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Item"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Items" ],
        "summary" : "Create item",
        "description" : "Create item",
        "operationId" : "createItem",
        "parameters" : [ {
          "in" : "body",
          "name" : "payload",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Item"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Created item",
            "schema" : {
              "$ref" : "#/definitions/Item"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/items/{itemId}" : {
      "get" : {
        "tags" : [ "Items" ],
        "summary" : "Finds an item",
        "description" : "Find item",
        "operationId" : "findItem",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "description" : "Item id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "An array of items",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Item"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Items" ],
        "summary" : "updates an item",
        "description" : "Update item",
        "operationId" : "updateItem",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "description" : "Item id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "payload",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Item"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Updated item",
            "schema" : {
              "$ref" : "#/definitions/Item"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Items" ],
        "summary" : "deletes an item",
        "description" : "Delete item",
        "operationId" : "deleteItem",
        "parameters" : [ {
          "name" : "itemId",
          "in" : "path",
          "description" : "Item id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "204" : {
            "description" : "Empty resonse indicating a succesfull removal"
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "LocalizedValue" : {
      "type" : "object",
      "required" : [ "language", "type", "value" ],
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
...
    "Price" : {
      "type" : "object",
      "properties" : {
        "price" : {
          "type" : "string",
          "description" : "price with two decimal places. e.g. 13.00"
        },
        "currency" : {
          "type" : "string",
          "description" : "Currency code (ISO 4217)"
        }
      }
    },
...
    "Image" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Image id"
        },
        "url" : {
          "type" : "string",
          "description" : "Image URL"
        },
        "type" : {
          "type" : "string",
          "description" : "Image type"
        }
      }
    },
    "Item" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Item id"
        },
        "title" : {
          "type" : "array",
          "description" : "Title of the item, multilingual",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "description" : {
          "type" : "array",
          "description" : "Description of the item, multilingual",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "categoryId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Category id."
        },
        "slug" : {
          "type" : "string",
          "description" : "Item slug"
        },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modifiedAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "expiresAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "images" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Image"
          }
        },
        "unitPrice" : {
          "$ref" : "#/definitions/Price"
        },
        "unit" : {
          "type" : "string"
        },
        "amount" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
  }
}